requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
websockets>=12.0

# LangChain & AI (optional for future LLM use)
langchain>=0.1.0
langgraph>=0.0.26
langchain-openai>=0.0.2
openai>=1.6.0

# Vector DB
chromadb>=0.4.18

# Code Analysis
radon>=6.0.1
gitpython>=3.1.40

# ML & Data Science
scikit-learn>=1.3.2
numpy>=1.26.0
joblib>=1.3.0

# Database
sqlalchemy>=2.0.23

# Utilities
aiofiles>=23.2.0
python-multipart>=0.0.6
httpx>=0.25.0

# Testing (CI / production QA)
pytest>=7.4.0
